test: extract AppStateHarness from TestSupport (#563, #434)
Eighth slice of splitting the large TestSupport.swift into domain files.
This is the biggest remaining chunk — the @MainActor struct AppStateHarness
that assembles a test-ready AppState by wiring together every domain mock.
Slices 434a-h shipped through #567.

Moves the @MainActor struct AppStateHarness (currently TestSupport.swift
lines 287-420, 134 lines, 6599 bytes — endpoint-verified before the
extract with sed -n '287p' == "@MainActor", '288p' == "struct
AppStateHarness {", '420p' == "}") byte-identical into a new
AppStateHarness.swift with the minimum imports (Foundation for
URL/UserDefaults/FileManager/UUID/TimeInterval/Data; @testable import
BugNarrator for AppState + all mock types). No XCTest — the struct
doesn't use XCTFail/XCTUnwrap/XCTestCase.

The @MainActor annotation is preserved on the line immediately above
struct AppStateHarness in the new file. No nested types, no fileprivate
dependencies, no additional isolation annotations inside the struct.

Byte-diff against origin/main lines 287-420 (per the byte-diff-against-
origin protocol, not against a working-state extract) is silent (exit 0)
— byte-identity proven. 662 unit tests pass unchanged; docs-sync
validation passes.

TestSupport.swift will not be empty after this slice — the remaining
content is the actor mocks (MockIssueExtractionService, MockExportService,
StubExportReceiptStore, filed as #557), plus makeSampleTranscriptSession,
waitUntil, and requestBodyData helpers that a future cleanup slice can
address.

Pre-build spec reviewed by codex in a fresh session (7 findings folded,
including corrected line range from stale 337-499 to actual 287-420 and
the byte-count checksum).
